_PDCLIB_scan: Had incorrect upper bound on field width causing integer overflow
git-svn-id: svn://rootdirectory.ddns.net/pdclib/branches/shepherd@485 bcf39385-58cc-4174-9fcf-14f50f90dd47

+++ functions/_PDCLIB/scan.c
@@ -162,7 +162,7 @@ const char * _PDCLIB_scan( const char * spec, struct _PDCLIB_status_t * status )
     status->width = (int)strtol( spec, (char**)&spec, 10 );
     if ( spec == prev_spec )
     {
-        status->width = SIZE_MAX;
+        status->width = INT_MAX;
     }
 
     /* Optional length modifier
